SEARC-2837 - NPE in AbstractAlternation.leftChild
The problem was that creating the clauses and putting them in the hash table for later reuse is not synchronized. This is manifested when the parent finder looks for a parent and the child has been created twice, leading to a false in equals since it's actually another object. 

Solution: 
Synchronize creation of clause's. We might want to make the synchronization less wide. I would also like a review of this.



git-svn-id: http://sesat.no/svn/sesat-kernel/trunk@6741 cdc3010f-a60b-0410-91f3-d90f869209c8