Issue SKER4943:  (IllegalAccessException "(Failed to visit object OrClauseImpl[(1 OR 2)])")
We do not need to set the setAccessible back to false for each call, since we are the only one that uses the method. And this will fix the error.

I also made the locking a bit better, so the normal operations will work without locking. Locking is only done when then unlocked version fails.


git-svn-id: http://sesat.no/svn/sesat-kernel/trunk@6754 cdc3010f-a60b-0410-91f3-d90f869209c8