Issue SKER4951:  (Encoding problems when characters like ,() are used in a query)
The problem was that the automatic detection of ISO-8859-1 failed, and the code thought that a string like ,ø was encoded in ISO-8859-1 and ended up as ,2 2Ã¸

I have tried to improve the detection, by inverting the process, compared to how it was done. I an not 100% sure about the .replacAll("[+]", "%20") and the same for *, but I think this might have been added to solve a similar problem before. So it would be good if you took a sec and looked at the diff.




git-svn-id: http://sesat.no/svn/sesat-kernel/trunk@6903 cdc3010f-a60b-0410-91f3-d90f869209c8