Issue SKER4947:  (StringChopper will not handle cdata)
Cleanup (braces, capital enums, non synchronized stack)

Some measurements:

all results are for 10x

http://localhost.no:8080/search/?q=oslo&x=40&y=15&c=m
New time is: 120
Old time is: 474 executed times:34308

http://localhost.no:8080/search/?q=brann&x=32&y=17&c=m

New time is: 127ms
Old time is: 284ms executed times:40276


http://localhost.no:8080/search/?q=brann&x=33&y=22&c=d
New time is: 25
Old time is: 83 executed times:8580


http://localhost.no:8080/katalog/brann/
New time is: 22
Old time is: 44 executed times:12090

This is measured by collecting the time spent inside the chop function. It is only done 10 times. Keep in mind that some of this might be done in parallel, so the result would probably be different when measuring it as a whole.

Still in a news search for oslo, we will probably see a significant overall speedup. 35ms for one search makes a difference.




git-svn-id: http://sesat.no/svn/sesat-kernel/trunk@6941 cdc3010f-a60b-0410-91f3-d90f869209c8