SEARCH-5262 Solr indexes to handle 2.18 query matching traffic
Reduce number of results we ask form from Integer.max to 25. If there are more results then 25, ask for the rest in a new query. This improves speed dramatically from 7000ms to 2ms in some cases. I haven't measured this compared to the fast predicates. We also used to get OutOfMemory for this queries, but this has gone now.


git-svn-id: http://sesat.no/svn/sesat-kernel/branches/2.18@7103 cdc3010f-a60b-0410-91f3-d90f869209c8